moved add btn code from loaned.js to a AddBtnLoaned component
